plotter/requirements.txt
datetime
python-dateutil
redis
jinja2
mpld3
matplotlib